+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/R.kt
@@ -1,6 +1,7 @@
 package cc.popkorn.example.model
 
 import cc.popkorn.annotations.*
+import cc.popkorn.core.Injector
 import cc.popkorn.core.model.Empty
 import cc.popkorn.core.model.Environment
 
@@ -11,7 +12,7 @@ class R1(d1: D1, d2: D2, d3: D3, d4: D4) : R1i
 class R2(d1: D5, d2: D6, d3: D7, d4: D8, d5: D9) : R2i
 
 @Injectable
-class R3(env: Environment, @WithEnvironment("env1") d1: DiA, @WithEnvironment("envX") d2: DiA, d3: DiA) : R3i
+class R3(env: Environment, @WithEnvironment("env1") d1: DiA, @WithEnvironment("envX") d2: DiA, d3: DiA, @WithEnvironment("env9") d4: DaJ) : R3i
 
 @Injectable
 class R4(ign: Empty, @WithEnvironment("env1") d1: DiB, @WithEnvironment("env2") d2: DiB, d3: DiB) : R4i
@@ -20,10 +21,10 @@ class R4(ign: Empty, @WithEnvironment("env1") d1: DiB, @WithEnvironment("env2") 
 class R5(str: String, int: Int, test: Float?, @WithEnvironment("env1") d1: DiC, @WithEnvironment("env2") d2: DiC, @WithEnvironment("env3") d3: DiC, d4: DiC) : R5i
 
 @Injectable
-class R6(@WithEnvironment("env4") d1: DiC, @WithEnvironment("env4") d2: DiC, @Alias("is7") d3: DiC, @Alias("is8") d4: DiC) : R6i
+class R6(@WithEnvironment("env4") d1: DiC, @WithEnvironment("env4") d2: DiC, @Alias("is7") d3: DiC, @Alias("is8") d4: DiC, d5: D12) : R6i
 
 @Injectable
-class R7(tmp: R0i?, d1: Wrapper.DiD, @WithEnvironment("envX") d2: Wrapper.DiD, d3: D10) : R7i
+class R7(injector: Injector, tmp: R0i?, d1: Wrapper.DiD, @WithEnvironment("envX") d2: Wrapper.DiD, d3: D10, d4: DaH) : R7i
 
 @Injectable
 @ForEnvironments("env1", "env2", "env3", "env4")