+++ popkorn/src/main/kotlin/cc/popkorn/annotations/Exclude.kt
@@ -1,7 +1,7 @@
 package cc.popkorn.annotations
 
 /**
- * This annotation applies ONLY for interfaces. It annotates the desired interfaces that
+ * This annotation applies ONLY for interfaces and abstract classes. It annotates the desired interfaces that
  * don't want to add as injectable. Needs to be BINARY to be detected between modules
  * when compiling
  *